# server/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go Fiber v2 기반 REST API 서버 - 스케치 기반 메뉴 추천",
        "title": "Ojeomneo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@woohalabs.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "api.woohalabs.com",
    "basePath": "/ojeomneo/v1",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "서버 및 데이터베이스 상태 상세 확인 (모니터링/디버깅용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "서버 상세 헬스체크",
                "responses": {
                    "200": {
                        "description": "상태 정보 반환 (DB 연결 실패해도 200)",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.HealthResponse"
                        }
                    }
                }
            }
        },
        "/healthcheck/live": {
            "get": {
                "description": "서버가 살아있는지 확인 (DB 연결 무관)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Liveness 체크 (Kubernetes startup/liveness probe용)",
                "responses": {
                    "200": {
                        "description": "서버 정상 가동 중",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.LivenessResponse"
                        }
                    }
                }
            }
        },
        "/healthcheck/ready": {
            "get": {
                "description": "서버가 트래픽을 받을 준비가 됐는지 확인 (DB 연결 포함)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Readiness 체크 (Kubernetes readiness probe용)",
                "responses": {
                    "200": {
                        "description": "트래픽 수신 준비 완료",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ReadinessResponse"
                        }
                    },
                    "503": {
                        "description": "트래픽 수신 불가 (DB 연결 실패)",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ReadinessResponse"
                        }
                    }
                }
            }
        },
        "/menus": {
            "get": {
                "description": "메뉴 목록을 조회합니다. 카테고리와 태그로 필터링할 수 있습니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "메뉴 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "카테고리 필터 (korean, chinese, japanese, western, asian, snack, cafe, other)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "태그 필터",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "페이지당 개수",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/menus/categories": {
            "get": {
                "description": "사용 가능한 메뉴 카테고리 목록을 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "카테고리 목록 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/menus/{id}": {
            "get": {
                "description": "메뉴 ID로 상세 정보를 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "메뉴 상세 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "메뉴 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sketch/analyze": {
            "post": {
                "description": "스케치 이미지를 분석하여 감정/분위기를 파악하고 어울리는 메뉴를 추천합니다",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sketch"
                ],
                "summary": "스케치 분석 및 메뉴 추천",
                "parameters": [
                    {
                        "type": "file",
                        "description": "스케치 이미지 (PNG/JPEG, max 5MB)",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "추가 텍스트 입력",
                        "name": "text",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "디바이스 식별자",
                        "name": "device_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sketch/history": {
            "get": {
                "description": "디바이스별 스케치 분석 히스토리를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sketch"
                ],
                "summary": "스케치 히스토리 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "디바이스 식별자",
                        "name": "device_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지 번호",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "페이지당 개수",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sketch/{id}": {
            "get": {
                "description": "스케치 ID로 상세 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sketch"
                ],
                "summary": "스케치 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "스케치 UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_handler.DatabaseStatus": {
            "description": "데이터베이스 연결 상태",
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean",
                    "example": true
                },
                "latency_ms": {
                    "type": "integer",
                    "example": 5
                },
                "message": {
                    "type": "string",
                    "example": "Database connection successful"
                }
            }
        },
        "internal_handler.HealthResponse": {
            "description": "서버 헬스체크 응답",
            "type": "object",
            "properties": {
                "database": {
                    "$ref": "#/definitions/internal_handler.DatabaseStatus"
                },
                "service": {
                    "type": "string",
                    "example": "ojeomneo-api"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "version": {
                    "type": "string",
                    "example": "1.0.0"
                }
            }
        },
        "internal_handler.LivenessResponse": {
            "description": "Kubernetes liveness/startup probe용 간단한 응답",
            "type": "object",
            "properties": {
                "service": {
                    "type": "string",
                    "example": "ojeomneo-api"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "internal_handler.ReadinessResponse": {
            "description": "Kubernetes readiness probe용 응답",
            "type": "object",
            "properties": {
                "database": {
                    "type": "boolean",
                    "example": true
                },
                "ready": {
                    "type": "boolean",
                    "example": true
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}